docstyle-pro_claude_v4/requirements.txt
python-docx>=1.1.0
lxml>=5.0.0
PyQt6>=6.0.0
Pillow>=10.0.0
